fix: 테스트 코드 분할
